feat: Support metapipeline
With the metapipeline, Prow will kick off a `PipelineRun` that itself
will generate another `Pipeline`/`PipelineRun`/etc. If the
metapipeline fails, the generated pipeline won't be created, so
behavior will be the same as it has been, but if the metapipeline
succeeds, we want to treat both the meta and generate pipelines as a
single Prow job, not marking the status of the Prow job as complete
until the generated pipeline has completed as well. The generated
pipeline's `PipelineRun` will exist before the completion of the meta
pipeline.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>